More sensible fix for Solaris build issue.

+++ dcraw.cc
@@ -2253,7 +2253,7 @@ void CLASS lossy_dng_load_raw()
       coeff[i] = getreal(12);
     for (i=0; i < 256; i++) {
       for (tot=j=0; j <= deg; j++)
-	tot += coeff[j] * pow(i/255.0, (int)j);
+	tot += coeff[j] * pow(i/255.0, (double) j);
       curve[c][i] = tot*0xffff;
     }
   }
